Changed data types to long long in GetSubstraction

+++ calculator-program.cpp
@@ -1,3 +1,3 @@
-int getSubtraction(int numberOne, int numberTwo){
+long long getSubtraction(long long numberOne, long long numberTwo){
     return numberOne - numberTwo;
 }
